Move auth.util.validate_current_user -> webscoket_api.ws_require_user
