+++ homeassistant/auth/mfa_modules/insecure_example.py
@@ -1,6 +1,6 @@
 """Example auth module."""
 import logging
-from typing import Any, Dict, Optional
+from typing import Any, Dict
 
 import voluptuous as vol
